# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a code scanner service API.",
        "title": "vibescan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/dashboard/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of past scans run by the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Scan History",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/armur-codescanner_internal_models.ScanHistory"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get overall statistics of scans for the logged-in user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "armur-codescanner_internal_models.ScanHistory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "critical_bugs": {
                    "type": "integer"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "high_bugs": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "low_bugs": {
                    "type": "integer"
                },
                "medium_bugs": {
                    "type": "integer"
                },
                "repository_url": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, running, success, failed",
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}